Removed "files" settings
After some discussion, it is better to store the file contents
in the actual context where they are used. So instead of listing a
bunch of generic blobs in "files" the file schema is now stored
in the settings that manages that file.
For example and interface now as an openvpnConfFile attribute for
openvpn interfaces.